# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kang Raling API Web Service",
...
        "contact": {
            "name": "Developer Website Kang Raling",
            "email": "danu3anugrah@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Server Lokal Development"
        },
        {
            "url": "https://kangraling.dlhgarut.id",
            "description": "Server Utama (Production Dinas Lingkungan Hidup)"
        }
    ],
    "paths": {
...
        "/galeris": {
            "get": {
                "tags": [
                    "Publik"
                ],
                "summary": "Ambil semua foto galeri",
                "description": "Mengembalikan daftar semua foto galeri untuk halaman publik. Tidak membutuhkan token. Hasil dipaginasi 12 foto per halaman.",
                "operationId": "d20cea3ff57aaf2d13b48bcfb1614d0d",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar foto galeri berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "gambar": {
                                                        "type": "string",
                                                        "example": "galeri/foto.jpg"
                                                    },
                                                    "slug": {
                                                        "type": "string",
                                                        "example": "kegiatan-bersih-sungai"
                                                    },
                                                    "keterangan": {
                                                        "type": "string",
                                                        "example": "Kegiatan bersih sungai Desa Sukamaju",
                                                        "nullable": true
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 24
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Galeri"
                ],
                "summary": "Upload foto galeri baru",
                "description": "Mengunggah foto galeri baru. Request HARUS menggunakan `multipart/form-data` karena ada upload file gambar.",
                "operationId": "5136fc67139353a7f1ff67c3d30eda55",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "gambar"
                                ],
                                "properties": {
                                    "gambar": {
                                        "description": "File gambar yang akan diunggah (jpg, png, dll)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "keterangan": {
                                        "type": "string",
                                        "example": "Kegiatan bersih sungai Desa Sukamaju",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Foto berhasil diunggah"
                    },
                    "401": {
                        "description": "Token tidak valid"
                    },
                    "403": {
                        "description": "Tidak punya permission kelola.galeri"
                    },
                    "422": {
                        "description": "Validasi gagal — file bukan gambar"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/galeris/{slug}": {
            "get": {
                "tags": [
                    "Publik"
                ],
                "summary": "Ambil detail foto galeri berdasarkan slug",
                "description": "Mengembalikan detail satu foto galeri berdasarkan slug. Tidak membutuhkan token.",
                "operationId": "89bbb246f8a964a2ce743408466c1546",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "kegiatan-bersih-sungai"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail foto galeri berhasil diambil"
                    },
                    "404": {
                        "description": "Foto tidak ditemukan"
                    }
                }
            }
        },
        "/galeris/{id}": {
            "post": {
                "tags": [
                    "Galeri"
                ],
                "summary": "Update foto galeri (gunakan _method=PUT)",
                "description": "Mengupdate keterangan atau mengganti foto galeri. Gunakan `multipart/form-data` dengan field `_method: PUT`.",
                "operationId": "83d6e73fa7495ba48d9c2a643ea8dcb2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "_method"
                                ],
                                "properties": {
                                    "_method": {
                                        "description": "Wajib diisi PUT untuk method spoofing",
                                        "type": "string",
                                        "example": "PUT"
                                    },
                                    "gambar": {
                                        "description": "Kosongkan jika tidak ingin ganti foto",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "keterangan": {
                                        "type": "string",
                                        "example": "Kegiatan bersih sungai (diperbarui)",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Foto berhasil diperbarui"
                    },
                    "401": {
                        "description": "Token tidak valid"
                    },
                    "403": {
                        "description": "Akses ditolak"
                    },
                    "404": {
                        "description": "Foto tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Galeri"
                ],
                "summary": "Hapus foto galeri",
                "description": "Menghapus foto galeri secara permanen. Fasilitator hanya bisa hapus foto miliknya sendiri.",
                "operationId": "1becc80d40edf2c37eec0210cbbe23f0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Foto berhasil dihapus"
                    },
                    "401": {
                        "description": "Token tidak valid"
                    },
                    "403": {
                        "description": "Akses ditolak"
                    },
                    "404": {
                        "description": "Foto tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token dengan format: **Bearer {token}**",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}